@@ -3,14 +3,14 @@
 # found in the LICENSE file.
 
 name: process_runner_example
-version: 1.0.0
+version: 2.0.0
 description: A an example for process_runner.
 homepage: https://github.com/google/process_runner/example
+publish_to: none
 
 dependencies:
-  args: ^1.6.0
   process_runner:
     path: ..
 
 environment:
-  sdk: '>=2.3.0 <3.0.0'
+  sdk: '>=2.12.0 <3.0.0'
