@@ -3,21 +3,20 @@
 # found in the LICENSE file.
 
 name: process_runner
-version: 3.1.1
+version: 4.0.0
 description: A process invocation astraction for Dart that manages a multiprocess queue.
 homepage: https://github.com/google/process_runner
 
 dependencies:
-  async: ^2.4.2
-  file: ^5.0.0
-  meta: ^1.1.2
-  path: ^1.5.1
-  platform: ^2.2.0
-  process: ^3.0.13
+  async: ^2.5.0
+  file: ^6.1.0
+  meta: ^1.3.0
+  path: ^1.8.0
+  platform: ^3.0.0
+  process: ^4.1.0
 
 dev_dependencies:
-  mockito: ^4.1.1
-  test: ^1.0.0
+  test: ^1.16.8
 
 environment:
-  sdk: '>=2.3.0 <3.0.0'
+  sdk: '>=2.12.0 <3.0.0'
