@@ -15,7 +15,7 @@ test_package.TypeMatcher<T> isInstanceOf<T>() => isA<T>();
 class FakeInvocationRecord {
   FakeInvocationRecord(this.invocation, [this.workingDirectory]);
   final List<String> invocation;
-  final String workingDirectory;
+  final String? workingDirectory;
 }
 
 /// A mock that can be used to fake a process manager that runs commands
@@ -66,8 +66,8 @@ class FakeProcessManager implements ProcessManager {
 
   ProcessResult _popResult(FakeInvocationRecord command) {
     expect(fakeResults, isNotEmpty);
-    List<ProcessResult> foundResult;
-    FakeInvocationRecord foundCommand;
+    late List<ProcessResult> foundResult;
+    late FakeInvocationRecord foundCommand;
     for (final FakeInvocationRecord fakeCommand in fakeResults.keys) {
       if (fakeCommand.invocation.length != command.invocation.length) {
         continue;
@@ -80,7 +80,7 @@ class FakeProcessManager implements ProcessManager {
         }
       }
       if (listsIdentical) {
-        foundResult = fakeResults[fakeCommand];
+        foundResult = fakeResults[fakeCommand]!;
         foundCommand = fakeCommand;
         break;
       }
@@ -93,26 +93,26 @@ class FakeProcessManager implements ProcessManager {
   FakeProcess _popProcess(FakeInvocationRecord command) =>
       FakeProcess(_popResult(command), stdinResults);
 
-  Future<Process> _nextProcess(List<String> invocation, String workingDirectory) async {
+  Future<Process> _nextProcess(List<String> invocation, String? workingDirectory) async {
     final FakeInvocationRecord record = FakeInvocationRecord(invocation, workingDirectory);
     invocations.add(record);
     return Future<Process>.value(_popProcess(record));
   }
 
-  ProcessResult _nextResultSync(List<String> invocation, String workingDirectory) {
+  ProcessResult _nextResultSync(List<String> invocation, String? workingDirectory) {
     final FakeInvocationRecord record = FakeInvocationRecord(invocation, workingDirectory);
     invocations.add(record);
     return _popResult(record);
   }
 
-  Future<ProcessResult> _nextResult(List<String> invocation, String workingDirectory) async {
+  Future<ProcessResult> _nextResult(List<String> invocation, String? workingDirectory) async {
     final FakeInvocationRecord record = FakeInvocationRecord(invocation, workingDirectory);
     invocations.add(record);
     return Future<ProcessResult>.value(_popResult(record));
   }
 
   @override
-  bool canRun(dynamic executable, {String workingDirectory}) {
+  bool canRun(dynamic executable, {String? workingDirectory}) {
     return true;
   }
 
@@ -124,8 +124,8 @@ class FakeProcessManager implements ProcessManager {
   @override
   Future<ProcessResult> run(
     List<dynamic> command, {
-    String workingDirectory,
-    Map<String, String> environment,
+    String? workingDirectory,
+    Map<String, String>? environment,
     bool includeParentEnvironment = true,
     bool runInShell = false,
     Encoding stdoutEncoding = systemEncoding,
@@ -140,8 +140,8 @@ class FakeProcessManager implements ProcessManager {
   @override
   ProcessResult runSync(
     List<dynamic> command, {
-    String workingDirectory,
-    Map<String, String> environment,
+    String? workingDirectory,
+    Map<String, String>? environment,
     bool includeParentEnvironment = true,
     bool runInShell = false,
     Encoding stdoutEncoding = systemEncoding,
@@ -156,8 +156,8 @@ class FakeProcessManager implements ProcessManager {
   @override
   Future<Process> start(
     List<dynamic> command, {
-    String workingDirectory,
-    Map<String, String> environment,
+    String? workingDirectory,
+    Map<String, String>? environment,
     bool includeParentEnvironment = true,
     bool runInShell = false,
     ProcessStartMode mode = ProcessStartMode.normal,
